Introduce openshift-on-ovirt
This suite will install and deploy openshift okd (3.10) on ovirt 4.2

Motivation
We want to establish working CI for the ovirt-openshift-extensions[1]
project, with its flexvolume implementation as top priority.

Modification
This test was driven by the reference archetecture work of openshift
RHV and adapted for the ci constrainst and interest.
It uses:
- ovirt orb for ovirt infrastructure in a click
- a container[2] which orchestrates all the installation
  and verification, from createing a VM in ovirt, installating origin
  master+etcd+node and deploying flex volume on top.
A python test will add basic sanity for spinning containers
using flex volumes which that will creates disks in ovirt.

Result
Running CI for ovirt-openshift-extensions

The suite will use oVirt Orb images in order to save time and avoid
false positives.

[1] https://github.com/oVirt/ovirt-openshift-extensions
[2] docker pull rgolangh/ovirt-openshift-extensions-ci

Change-Id: I3feae85953e6770746c034a7c89cd918f9ab22bf
Signed-off-by: Roy Golan <rgolan@redhat.com>
Signed-off-by: gbenhaim <galbh2@gmail.com>