network: React on succesful pinging the VM
test_connect_vm_to_external_physnet succeeds, if the VM
replies to the ping. This way the test does not have to
wait until the VM is marked as UP.
Excluding the time to create the fixtures will be required
by other tests, this tests reduces the execution time of
test_connect_vm_to_external_physnet down to around 30 seconds.

Change-Id: Ie2b1d47bf52a9cb173d02a3a68449ee8f75c2980
Signed-off-by: Dominik Holler <dholler@redhat.com>