networking: Introducing test_hot_unlink_vnic
This test will verify that marking a vnic of a running vm as unlinked
actually unlinks it with no errors.

Currently, due to a bug in RFE https://bugzilla.redhat.com/1451342
(configure guest MTU based on underlying network), the test fails.

Once the bug will be fix, the 'xfail' should be removed from the test.

Change-Id: Iccf661fd3d43b9928fe9ac70fe98aaa09b45d42d
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>