Adjust "list_glance_images" error message
The test name was changed in ff71fb4fdc0797a60f015919e68ec3d371420a08,
thus the error message should be adjusted as well.

It will be better not to refer to the function name from inside the
function, in order to avoid duplicate changes in the future.

Change-Id: I698d8327a95b54491ef821bbcbe983323ce96424
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2375
Signed-off-by: gbenhaim <galbh2@gmail.com>