network: syncutil: Introducing error_criteria
In many cases, exec_func may raise errors. These errors may or may
not be transient.

Up until now, we've left the responsibility to decide whether an error
is fatal or not to exec_func -- however, this leads to convulted code
and adds responsibility to exec_func.

After trying to move error handling to success_criteria I am now thinking
it would be best after all to introduce a new optional function
to evaluate errors.

In the majority of cases, this is not needed -- any exception will
be deemed fatal. However, in cases where a temporary error may occur,
users may pass a function that handles the error and determines
whether to continue retrying, or explode.

Another case is where we don't need to evaluate the result, e.g.
where we want to initiate a command and only interested in it not
failing (e.g. retrying on errors such as simultaneous running tasks)

As such, I've made success_criteria optional as well.

Change-Id: I0db3fa61522c6077e4413edcbb241520682e2640
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>