Repos: add python2-ovsdbapp to the reposync config
This is a temporary workaround to update the ovsdbapp version
from 0.6.0 to 0.12.1, which is required by ovirt-provider-ovn.

Introducing security group support to the ovirt-provider-ovn
project requires ovsdbapp > 0.12, which adds port group table
support.

Unfortunately, the version of ovsdbapp currently released in cbs
 - 0.12.0 - adds plenty of openstack library dependencies that
oVirt does not need or is interested in importing.

ovsdbapp-0.12.1 release removes those library dependencies, and
when that version is released by openstack on cbs, we should just
link to it, and remove this temporary repo fix.

Change-Id: I54d4c3edff6c8c1773aa9ccd43453e0de1c3c25a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>