Physnet test w/ port security and security groups
During the physnet test, the VM port is updated, enabling its
ovn port security attribute.

As a result of this update, all ingress IP traffic to the VM is
dropped, which means the ping command to that VM will fail.

Afterwards, a security group rule is added - through ansible - to
the default security group - which is attached to the VM port -
allowing ingress ICMP traffic from the pinging host IP.

When that rule is provisioned, the ping command succeeds.

Change-Id: Iabd17b3f4c353155e6a3eebe3a3fd60f6633eec2
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>