networking: introduce Node with ssh capabilities
The hostlib.Host and virtlib.Vm objects only represent the underlying
oVirt entities which are limited to exposing actions via the sdk and
do not permit other actions such as ssh-ing into the actual machines
that these objects represent.
The newly introduced Node object will provide ssh access to the actual
machines and will expose different actions that will be required for
tests.

Change-Id: If24acb7fe090655d3be86da27a7afa1ff52e8c2a
Signed-off-by: Eitan Raviv <eraviv@redhat.com>