Fix notifier check in master and 4.2
Up until now we have checked for the presence of VDC_STOP event in
/var/log/messages to find out if ovirt-engine-notifier is working
correctly. Unfortunately this was not reliable, because this event is
raised only when engine is not responding for at least 90 seconds, and
we have performed only standard restart of ovirt-engine service, so on
faster machines this check would fail.

This patch changes the event to USER_VDC_LOGIN, which is raised always
when a user successfully logs in into engine.

Change-Id: I0ae410a80f448538fe89e1acfe2f4e65041297df
Signed-off-by: Martin Perina <mperina@redhat.com>