basic-suite-master: Convert network name to utf-8
networks_service() seems to return a str object, not a unicode one.

This causes detach_vm_network_from_host_0 to emit:

Unicode equal comparison failed to convert both arguments to Unicode -
interpreting them as being unequal

It also causes the assert_false there to be useless, because it will
never find this network, whether (still) connected or not.

Not sure the fix should be here and not in the SDK.

Change-Id: I35232f46d48592863167a428758ad70f58dc19ab
Signed-off-by: Yedidyah Bar David <didi@redhat.com>