upgrade-from-release-suite-master: Fix DC/cluster version
DC/cluster versions were common for all upgrade suites, but each upgrade
suite should have its own DC/cluster version which is created after
upgrade and its own list of DC/cluster versions which DC/cluster should
be upgraded to. Specifically this change will create DC/cluster 4.1
(previously 4.0) and afterwards it will try to upgrade it to 4.2 and
then 4.3 (previously to 4.1 and 4.2).

This change is required because in 4.3 we are going to drop DC/cluster
version 4.0

Change-Id: I40f7172f042379eab5cb9f293838dd2d25139e3b
Signed-off-by: Martin Perina <mperina@redhat.com>