vagrant: Adding reposetup based on a container
the mkrepo container contains the logic of "lago ovirt reposetup",
which creates the test repo.

Instead to spawn the reposerver using python, we use an nginx container.
The nginx container will be stopped as soon as run_suite_vagrant.sh
exists.

reposync cache will be stored in /var/host_cache if exists, otherwise in
the .cache directory in the user's home dir.

Also, updating the reposync config with new packages and excluding
the appliance and node.

For now keeping the external repos open, we should close them once we
adjust the reposync builder to work with vagrant.

Change-Id: I333ce9ddde7c927d282cd4a708d434e1ad0c1a97
Signed-off-by: gbenhaim <galbh2@gmail.com>