Remove screenshot directory before creating a new one
There is no code outside automation that removes exported-artifacts.
When UI tests are run, they create a screenshot directory in
exported-artifacts.  If the directory already exists, perhaps from a
previous OST run, then OST fails late in the testing, without
providing an intelligible explanation.

This patch removes the previous content of the screenshot directory
automatically if the directory already exists,

Change-Id: If78727edc904c761f6153ef44bf0346425588945
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>