upgrade-from-release-suite-master: Create template and subtemplate
Create template and subtemplate from VM with interface before
upgrade. This flow should test issues that were caused
by vm_id and vmt_id that is present in the vm_interface table [1].

The bootstrap flow was moved before the update
so we can create the vm and its templates.

Run VM after the engine upgrade is finished
to check that VM is still runnable with older
vdsm. Then upgrade the hosts to continue with
the tests from the same state as before.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1547038

Change-Id: I504a09b6ad81876b2f868f55233a2111751b0151
Signed-off-by: Ales Musil <amusil@redhat.com>