Wait for host to be up - after x retries
The problem is actually that hosts can switch
from Up to Initializing.
we need this patch to handle that.

We assumed that at least one of the hosts is already up when
initializing a storage domain.
This assumption is actually not always true.
Let's wait, with a timeout, for some host to come up rather than raising
an exception about missing hosts immediately.

The idea to check for Up status again and hope that the host won't switch
to another status before it is used for the given purpose

Change-Id: I8041995aded050fa03e21f8be5d6bdf50f9ff155
Signed-off-by: Galit Rosenthal <grosenth@redhat.com>