ovirtlib: log engine event on every test start
With the increased number of tests and their complexity it is sometimes
difficult to correlate  a failure in the engine log with the
corresponding step in the test.

To help delineate the start and end of each test, a fixture is
introduced which adds an event to the engine log whenever a test
function is first called.

Change-Id: Ie90d19fc1e978b32b62f1342451323b3fc149b33
Signed-off-by: Eitan Raviv <eraviv@redhat.com>