networking: host_0 dynamic ipv6 assignment
In order to allow communication over ipv6 subnets among lago hosts and
engine an actual ipv6 address is required on the NICs of these machines.
When lago setup is complete the NICs have autoconf configured but no
ipv6 address assigned. This patch introduces fixtures that assign ipv6
addresses to the NICs.

Change-Id: Ic0a8bd97b3a54d12fb880135a8e022e0ba38e164
Signed-off-by: Eitan Raviv <eraviv@redhat.com>