basic_suite: Install VDSM SOS plugin if missing
VDSM SOS plugin has been from VDSM to SOS project, but unfortunately
there is a transition period between the time where package was removed
from VDSM and it's provided by new SOS release. So if the VDSM SOS
plugin doesn't exists on hosts, we are going to install it manually.

Change-Id: I5fa52197e302c1c8ef5827f77e73fee54aaf8225
Signed-off-by: Martin Perina <mperina@redhat.com>