basic_suite_4.3: Install VDSM SOS plugin if missing
VDSM SOS plugin has been from VDSM to SOS project, but unfortunately
there is a transition period between the time where package was removed
from VDSM and it's provided by new SOS release. So if the VDSM SOS
plugin doesn't exists on hosts, we are going to install it manually.

Change-Id: I0744239eaae50000ebc984994ed49d5d7d50e96e
Signed-off-by: Martin Perina <mperina@redhat.com>