network, ovirtlib: Expose the root system through sdkentity
The oVirt SDK entities are represented through several interfaces that
are defined in the sdkentity.
The entities can be split into two groups: The ones which are
dependent on the system root and the ones which are dependent on other
entities.

New scenarios have been appearing which require non-root entities to use
the system root internally.
In order to support such usages, the root system is exposed in all
entities, root or otherwise.

Change-Id: Id7f43f2945677a7f65130d13592a5e5e2a7693ed
Signed-off-by: Edward Haas <edwardh@redhat.com>