network-suite-master: test injecting sec groups via vnic profile
Add test that assigns security group as custom property of vnic profile
and checks whether the security group applies to the port. Add cirros
to vm, because to update the port, hotunplug is required and that
requires vm with os. Also move the custom property injection from
test_modify_vnic as it's more logical to do in this new test and rename
it accordingly to test_assign_network_filter.

Change-Id: I113af0353cca33bdfb5d46d8c5d30ee9d759025c
Signed-off-by: Andrej Cernek <acernek@redhat.com>