Prepare sharing libguestfs download across suites
To modify a guest image with a XFS version newer than CentOS7
can handle, the libguestfs appliance has to be downloaded
manually, please find details in bug 1629845.

The function install_libguestfs downloads a CentOS8 compatible
libguestfs appliance and sets the related enviroment variable
to let the libguestfs based tools use the new appliance.

This function is moved to run_suite.sh, because it will be
required for all CentOS8 based test suites.

Change-Id: I8e2310e4e448dc3c3a6c41fa2558eeed817aa0aa
Signed-off-by: Dominik Holler <dholler@redhat.com>