basic-suite: Use /etc/hosts
Lago configures an IPv6 and an IPv4 resolution of hostnames.
But the configured IPv6 address does not match the address
which is provided to ovirtmgmt interfaces.
During the initialization of the migration, the CentOS 8 host
tries to reach the migration destination host via the IPv6
address configured by lago, but not the one which is used on
ovirtmgmt. This why the migration would fail.
To enable the migration test again, the host names are
resolved by /etc/hosts, to avoid the problem.

Change-Id: I54aebe17614026dbccd94a84f93754f75eb02c41
Signed-off-by: Dominik Holler <dholler@redhat.com>