Added "sar" system resources collection on VMs.
"sar" is a simple utility available in CentOS base. This change adds a deploy
script that starts "sar" in background and asks to write system statistics
to /var/log/sarstats.bin. By doing that we try to alter the system as less
as possible, just one standard base package and a background process.

Since sometimes the lago VMs are not automatically destroyed (e.g. during
local runs), we also start a background sleep followed by a kill that will
terminate "sar" in given time (2 hours currently). With this setting the
resulting file is not longer than 100 Mb.

Statistics is optional in terms of testing and we make sure that in case
something is not going ok in the deploy script, we do not fail the tests.

Change-Id: Iaaf6cfefa32decb4ea4a574a4ab52c01d0449056
Signed-off-by: Anton Marchukov <amarchuk@redhat.com>