basic suite: Don't use virtualenv
Basic suite uses virtualenv to install selenium because there
is no RPM version available. The problem is, that the created virtualenv
is used in a very ugly and hacky way.

Since the tests are run with 'env_run_test' helper [1][2], which itself
refers to lago-ost-plugin [3], which finally uses nose underneath [4],
the virtualenv gets lost somewhere in that jungle. To enforce
the visibility of the installed selenium package, we modify
the 'PYTHONPATH' variable [5].

This patch proposes a different solution - instead of using virtualenv,
it installs selenium with 'pip --user', causing the package to land
in '$HOME/.local/lib/...', with '$HOME' itself being the root of the
git repo in the particular case of mock. That way it's visible without
additional hacks and gets cleaned up nicely.

[1] https://github.com/oVirt/ovirt-system-tests/blob/a3656887d92e379bad012f6a13d1b01857c9112a/basic-suite-master/control.sh#L39
[2] https://github.com/oVirt/ovirt-system-tests/blob/a3656887d92e379bad012f6a13d1b01857c9112a/run_suite.sh#L274
[3] https://github.com/oVirt/ovirt-system-tests/blob/a3656887d92e379bad012f6a13d1b01857c9112a/run_suite.sh#L279
[4] https://github.com/lago-project/lago-ost-plugin/blob/2d060db1389202771d43cc2f4648868766d646be/ovirtlago/prefix.py#L271
[5] https://github.com/oVirt/ovirt-system-tests/blob/a3656887d92e379bad012f6a13d1b01857c9112a/basic-suite-master/control.sh#L35

Change-Id: I223e1046ce82837774534cb677246dc17dda9fa7
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>