el8: Don't try to collect whole '/etc/httpd' dir
When running basic suite on CentOS 8 we get failures on artifact
collection of '/etc/httpd' directory from the engine. Turns out,
that inside this dir there's a symlink '/etc/httpd/run' pointing
to '/run/httpd', which contain all the httpd's sockets. This causes
the collection process to fail.

To fix that, we're simply collecting interesting subdirs of '/etc/httpd'
directory.

Change-Id: Iedee40881822000cc577429a9c90858587efbbe0
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>