wait longer between retries for host status
wait_for_status was supposed to resolve flaky status reporting of host
states, but the interval is only 1s. In reality the flapping is slower
than that, let's go with 10s instead

From one of the runs:
05:11:36,456 Status of host lago-basic-suite-master-host-0 was set to Up
05:11:40,929 add_master_storage_domain falure on "host not up"
05:11:44,803 Status of host lago-basic-suite-master-host-0 was set to Up
the state was apparently wrong for some time in between. Not sure if 10s
is going to be enough...but we won't know until we try

Change-Id: Ia109b6b1b2a2ddb1cdfd7be7c3fcd0d4bee840af
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>