pytest: Port '100_basic_ui_sanity' test module to pytest
This patch rewrites the ui sanity tests to make them work with pytest.

The tests are executed in the order of their definitions in the module.

The new implementation uses the much more flexible 'hub_url' fixture
to get a working selenium grid instead of setting things up by itself.

The attempt to add the engine's root CA to browsers has been dropped
completely, and all capabilities are set to accept insecure
certificates. This change was made because:

- it doesn't work currently
- the way to add a trusted root CA is browser-dependent
- the way to add a trusted root CA is also container backend-dependent
  and practically impossible to do automatically when using an existing,
  external selenium grid
- there is no nice, unified selenium API to do that

Instead of trying to add a trusted root CA to selenium browsers, two
simple, curl-based tests have been added to ensure that the
engine-provided root CA can be used for secure connections.

The image-uploading test has been changed not to generate its own
"dummy_disk" (because it's container-backend dependent and impossible
to do on an external grid), but to use '/etc/hostname' file as the
image payload. This should work on any possible machine. The alias
for the image has also been changed from container name to browser
name + current time in seconds (so that multiple runs of that test
won't end up in any 'image already exists' conflicts).

Change-Id: I32b411f320c9ed66145f9c867f112b34841bc1ce
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>