pytest: Add 'repo_server' fixture
This patch adds a 'repo_server' fixture that runs an HTTP server
with the packages for dnf. The implementation of the server is currently
provided by lago-ost-plugin, but in the future we will try to use
a different solution (i.e. container).

The fixture is set to 'autouse' and added to 'conftest.py' meaning every
test will use it automatically.

Change-Id: I356a9e447e260b440e9ce7769975a9c19085ee02
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>