poll update disk
To make the wait for hotunplug completion reliable, this patch
adds polling on the correlation id. Since engine excutes UpdateDisk
when hot-unplugging (REST only), we need to know when it's finished.
UpdateDisk also locks the disk, so this patch also waits for the
disk to be unlocked (should only remain lock in case of a failure).

Change-Id: I563c53b7bc55169a37369a0262c4b3259af8bd16
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>