pytest: Restore '002_bootstrap_pytest' module
Now that storage setup has been moved to deployment phase and the issues
around unavailability of alocalsync repository have been resolved with
'repo_server' fixture we can restore to pytest version of bootstrap
module.

Previously the ordering was done with 'pytest.mark.run(order=)'
decorators, but that proved inserting tests in between other ones and
disabling single tests to be difficult. Now we use 'order_by' decorator
which is controlled by the list of test names to be run.

Change-Id: I1b3b52709628af2bcc475e83e5bd6fdee2ad2275
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>