selenium: podman: Small improvements
- Use force when removing podman pod

From time to time some issues arise when cleaning up
the podman's container environment used for selenium.
This patch adds a flag to force the removal of the pod
on cleanup.

- Allow running multiple grids

This patch makes running multiple grids that use podman as a backend
possible. It uses simple probing to find first free port that the
grid could use and passes that to the selenium containers.

Change-Id: I83496c4939297644efbe4887d48016b80b8f58bb
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>