selenium: podman: Add grid setup retries
From time to time podman-based selenium grid fails to start up properly.
The logs indicate that both the hub and the nodes seem to be fine
on themselves, but the communication between them is not working.

This patch adds a global-level retry for the grid - after a failed
attempt, the whole thing is torn down and set up once again.
It also adds pod inspection logs for a failed attempt.

Change-Id: I5029631675c4b50130219033c2786828bfa47be5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>