ost-images: Add proper RPM versioning
This patch makes the built RPMs to have proper versions.

The versioning scheme is very simple:
- the major version is taken from the 'ost-images.spec.in' template
  and is currently equal to '1'. It should be bumped only when the base
  image changes.
- the release is dynamically generated and timestamp-based i.e. '20200507102030'
- together they form the full version i.e. '1-20200507102030'

The "upgrade" RPM can have a dependency to the "base" RPM of either:

 Requires: %{name}-%{distro}-base = %{version}

for iso-based installations and:

 Requires: %{name}-%{distro}-base = %{version}-%{release}

for repo-based installations. The difference here is that for iso-based
installations the base image should not change very often and when it does,
we should bump the major version.

For repo-based installations however it's the "base" layer that carries
the upgraded packages and it's built nightly, so we need the more strict
dependency.

The "*-installed" RPMs have dependencies to the upgrade layer of:

 Requires: %{name}-%{distro}-upgrade = %{version}-%{release}

so they work only with the exact match of the "upgrade" RPM.

Change-Id: Id18e035876d51c7db07c85361e9da5f0631e43dd
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>