setup: Do not hide real issue by reserving ports
ovirt-imageio ports (54322, 54323) were reserved on hosts and engine to
avoid failure to startup the service when a port is used by some other
process.

This "fixes" the issue by hiding it, so OST will not fail when we have a
real issue in the product that OST tries to test.

Port reservation must be handle by the product:

- On the host this was fixed in 2018 in:
  https://gerrit.ovirt.org/c/96447

- On engine this was fixed today in:
  https://gerrit.ovirt.org/c/108943

Remove port reservation from all setup scripts.

Change-Id: I6a435f886eefcb4aba4d8ca59d2e458b98e5435c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>