ost-images: Use OST images when available
This patch switches el8-based basic and network suites
to use prebuilt VM images [1]. El7 runs are unaffected.

Since the images already have host- and engine-related
RPMs installed, we need to add a simple 'dnf upgrade' run
before the tests kick in. This is needed in case someone provides
extra repos with custom-built versions of those RPMs.
This is a temporary workaround - ultimately we want to build
a VM image with custom RPMs before OST run automatically
and get rid of all the deployment scripts.

When running inside mock, we have to make the qcows visible by both
libvirt, which is running outside of mock, and lago, which is running
inside the mock. This is why we create an 'ost-images-XXXXXX' directory
in '/var/lib/lago/' and make copies of images there.

[1] https://templates.ovirt.org/yum/

Change-Id: Ie64a2e26504d3252220178262ff22b6ccb9e074c
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>