ost-images: Always build base image for repo-based installations
Until now, the decision whether to build the base image or not was
solely based on the presence of the package on the build machine - if you
had 'ost-images-{distro}-base' package installed, we used the installed
version to build other layers.

This is not correct for repo-based installations - in this case the base layer
actually performs the function of both base and upgrade layers, so we should
always rebuild the base image. This patch implements that behavior.

Change-Id: I52ae6af71f746a844333733636357fdfa08dec66
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>