lago: Skip sysprep when ssh key is available
Since we've added ssh keys to 'base' layer of ost-images [1],
and support for using existing ssh keys was added to lago [2],
we no longer have to go through 'sysprep' stage. This saves
a lot of time since 'sysprep' was doing selinux relabeling,
which is required after adding ssh keys for root user.

[1] https://gerrit.ovirt.org/109955
[2] https://github.com/lago-project/lago/pull/813

Change-Id: I5132971b94ac1ba5c107636dc48b24d11f5ebc15
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>