network: test pool allocation finds unused macs
Adding a vnic to a VM when there are still available MAC addresses in
the MAC pool fails as per the scenario BZ 1760170. The failure occurs
because the search pointer for the next available address in the first
pool is reset to the beginning of the pool after creating 2 vnics on
VM1. This is because the pool has only two addresses to begin with.
The first test reproduces this scenario. The second test shows that if
the first pool was not exhausted prior to moving VM1 to a different
cluster, the failure does not reproduce.

Engine has been fixed so that a pool will search for and allocate all
free and unused addresses thereby fixing the BZ scenario.

Change-Id: If60d90af180734eccdd1a13e3e311ae27eb8bd35
Bug-Url: https://bugzilla.redhat.com/1760170
Signed-off-by: Eitan Raviv <eraviv@redhat.com>