network: retry setup_network for suite < 4.4
Reintroduce the retry to avoid recurring failures on suite 4.3 until
this suite is not longer run by CI.

The retry was initially introduced in gerrit.ovirt.org/#/c/93978 when
ovirt-engine policy was reject-on-busy. For versions 4.4 and up, the
policy has been changed to queue-on-busy so the retry is no longer
needed.

Change-Id: Ibf6cbccf77a53e6306fdf86bfa2e5ea5af32c0f1
Signed-off-by: Eitan Raviv <eraviv@redhat.com>