basic: Verify glance template creation before tearing down the engine
We import some images from glance in [1] and use them
for creating VMs. The verification of the import operation is done
here [2]. In between the import, and the verification we run an engine
backup and restore [3].

The problem with this scenario is, that while the disk are imported
successfully from glance, the template creation operation is dropped
(probably because of engine backup-restore). This causes failures
when trying to create 'vm1' from the template later on to happen [4].

I think the reason why this worked until now is that
'test_verify_add_all_hosts' [5] was taking so much time, that
the template was already created successfully in the meantime.
With the introduction of pre-built VM images both hosts have 'vdsm'
installed already, so the flow is much quicker and exposes this flaw.

This patch moves the verification of glance imports to happen
before running engine backup-restore.

[1] https://github.com/oVirt/ovirt-system-tests/blob/7ece675ed85ddc669af46ab0efac0be974e24b93/basic-suite-master/test-scenarios/002_bootstrap_pytest.py#L606
[2] https://github.com/oVirt/ovirt-system-tests/blob/7ece675ed85ddc669af46ab0efac0be974e24b93/basic-suite-master/test-scenarios/004_basic_sanity_pytest.py#L1179
[3] https://github.com/oVirt/ovirt-system-tests/blob/7ece675ed85ddc669af46ab0efac0be974e24b93/basic-suite-master/test-scenarios/002_bootstrap_pytest.py#L1278
[4] https://jenkins.ovirt.org/job/ovirt-system-tests_standard-check-patch/10216/testReport/junit/basic-suite-master.test-scenarios/004_basic_sanity_pytest/test_add_vm1_from_template/
[5] https://github.com/oVirt/ovirt-system-tests/blob/7ece675ed85ddc669af46ab0efac0be974e24b93/basic-suite-master/test-scenarios/002_bootstrap_pytest.py#L436

Change-Id: Iaaea78406a3ca0adf5e312c161ad597cda7d4976
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>