ost_utils: ansible: Make ansible fixtures multithread-compatible
Getting ansible fixtures to work in multiple threads at once
is a bit tricky - the problem is the temporary directory that
ansible-runner uses for tracking state. Trying to use the same
directory for all threads ends up with ansible-runner reporting
a conflict error. OTOH we cannot create a new temp dir for each
use of ansible modules because we loose all gathered facts.

The solution is to have temporary directories created per-thread
and the process of fact-gathering to be run once for each thread.
This patch implements that.

On top of that, we track all temporary directories created in all
threads and remove them after the testing session is over, not
to leave a mess behind.

Change-Id: Iff3ad1c1a1fe2cdd5c9457a337f9fa183e83f2c8
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>