network: host change cluster - lib-erate + objectify
1. Replace fixture and hostlib function for changing host cluster with a
   contextualized method on hostlib.Host
   This serves several purposes:
   - better code fluency because functionality is on the object where it
     belongs
   - less dangling wrappers
   - better reuse

2. Modify 'host_1_on_cluster_0' to 'host_0_on_cluster_0' for symmetry
purposes in following patches. This does not impact existing tests
because host_1 is not a participant in the tests using this fixture.

3. The contextualized host.toggle_cluster() func is called from within a
local fixture in the test module, because the fixture provides separate
logging for the setup phase, thereby clarifying that a failure in its
functionality does not belong to the testing phase.

This patch is a preparation for more usages in next patch.

Change-Id: I83a4017c79dc18dc2a20372315325dde677431a1
Signed-off-by: Eitan Raviv <eraviv@redhat.com>