ost_utils: engine: Add 'engine_download' fixture
This patch introduces an 'engine_download' fixture that eases
making http(s) requests to the engine.

The fixture uses curl underneath and handles http and https
urls in different ways:
- if http is used, we add a mapping that allows to reach the engine
  with its fqdn ('engine' ATM of writing)
- if https is used, we use the 'engine_cert' fixture to obtain the CA
  from the engine and pass it to curl with '--cacert' flag. Similarly
  to http, we also add a mapping that allows reaching the engine
  with its fqdn.

Depending of the value of the 'path', 'engine_download' either saves
the downloaded contents to a file, or returns them as bytes.

Since we're passing the '-f' flag to curl, any 4** or 5** http
status code will cause failures.

Change-Id: I95cd3b5971e4e08da94ed88bc62c2eee0f249431
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>