basic: ansible: Port 'test_add_hosts' to ansible
This patch gets rid of lago-based 'prefix' in 'test_add_hosts'
in favor of ansible fixtures.

To make the port possible, a 'root_password' fixture has been
added with the password hard-coded. We should probably improve
that later to obtain the password from OST backend, but since
we don't change it that often, it works for now.

Change-Id: I6f68d3c70aeeb5db2b7db14e2c36eb0f243a0a17
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>