ost_utils: engine: 'engine_restart' fixture
We should avoid restarting the engine in OST in general.
However, we currently have 2 cases of this happening in basic suite.
The problem is both of these cases do a simple systemd unit stop-start.
This is not enough - the systemd unit being up doesn't mean the engine
is fully functional (actually getting 100% sure, that the engine is
fully working is not possible ATM). This causes problems with tests
that follow engine restarts - they expect everything to be working.

Since tests should be as atomic as possible, this patch adds
an 'engine_restart' fixture and replaces the aforementioned use
cases to use it. When called, it does a stop-start
cycle of the systemd unit and then waits 30 secs for the health
check endpoint to return 200. This is better than what we have
before, but still may need improvement in the future. Having
the code in a single location should make the improvements easier.

Change-Id: I8a68cc53f37fa286ef9824de004e0bd91c168468
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>