basic: ansible: Improve 'test_download_engine_certs'
This patch improves 'test_download_engine_cert' test and gets
rid of lago-specific stuff in favor of ansible fixtures.

The original test had a couple of issues:
- it used custom, complicated code to download stuff from the engine
- it didn't actually verify downloaded certificates
- it covered health-check endpoint, which is not really a certificate

The code has been split to 2 test cases:
- test_verify_engine_certs
- test_engine_health_status

The first test uses http to download the CA and ssh pubkey and verify
them with appropriate commands. If the verification fails, the contents
of the certificate/pubkey are printed to stdout.

The second test uses both http and https to access the health-check
endpoint. With the https check here, we can be pretty sure that
the certs are indeed fine.

Change-Id: I5780852663788ac8e4b24c5d6e8e731173804220
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>