basic: performance: decouple 'control.sh'
Until now 'control.sh' file was shared between performance suite
and basic suite. Since performance suite is not 100% ported to pytest
yet, the script had to support both running tests modules with pytest
and the legacy way. This patch decouples 'control.sh' by changing
the symlink to a copy so we can move on with basic suite.
When performance suite is ported to pytest this patch can be reverted.

Change-Id: Ie19edeea8c3a7f8741f8ba2ec44f7b1afd673847
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>