basic: Replace 'pytest-ordering' with custom ordering code
We use the 'pytest-ordering' plugin along with '@order_by' decorators
to define the order of the tests in bigger modules.

With [1] we're changing how we run test modules - instead of running
a separate pytest session for each of the modules, we run them in bulk.
This however messes up the ordering.

This patch replaces the 'pytest-ordering' plugin with a custom
'pytest_collection_modifyitems' hook that takes the names of the
test modules into account and makes the ordering work back again.

[1] https://gerrit.ovirt.org/#/c/111441

Change-Id: Ibb231ab405498b310db07ecaccbeef1c498cc9c4
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>