ost_utils: network_utils: Add 'get_ips[6]' functions
This patch adds 2 functions to 'ost_utils.network_utils' module:

- get_ips
- get_ips6

The functions return all ip v4/v6 addresses a VM has in a specific
network. The resolution is done using ansible and the only lago-dependent
part is the mapping of hostname --> network --> network ifaces obtained
with the 'ost_utils.backend' module.

The correctness is validated by rewriting the 'engine_ip' fixture
to use the new functions.

Change-Id: Id94b760e71e61211d7a478f0dcb3b4db01231793
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>