ost_utils: Move 'TestEvent' to 'ost_utils'
The 'TestEvent' context manager from the 'test_utils' module is quite
handy. Unfortunately 'test_utils' is only available in the basic suite.

This patch moves this utility to a new 'ost_utils.engine_utils' module,
so that everyone can use it. It does that with a new, more intuitive
'wait_for_event' name and gets rid of lago-provided 'testlib' module usage
on the way.

The moved context manager is reimported in the 'test_utils' module, so that
existing code does not break.

Change-Id: Ie4b495945490ef81a226d60ad4272d680f3a2c26
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>