basic: Make '_add_storage_domain' more independent
This patch drops usage of 'test_utils' and 'testlib' modules
in '_add_storage_domain' function. It also adds a 'dc_name' argument
to its signature. The goal here is to make the code more generic
and independet of any basic-suite-specific and lago-related
modules. The function will be extracted to a seperate module
in subsequent patches.

Change-Id: I34982c3a8933934dcb6279c18cff58dc08219efd
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>