ost_utils: ansible: Always return the result
Currently you can query ansible facts like this:

 assert ansible_engine_facts.get("ansible_hostname") == "basic-suite-master-engine"

This is a very nice feature that gives you direct access
to the queried values. When running ansible modules however:

 ansible_engine.shell("cat /etc/passwd")

the returned value is an instance of 'AnsibleRunner' class.
This is not very useful, so this patch changes that.
Now usage of ansible modules will give you the actual result
of the module, i.e. following the example for the 'slurp' module [1],
we can do this:

 content = base64.b64decode(ansible_engine.slurp(src="/etc/passwd")["content"])

This doesn't change the mechanics of errors - in this case an exception is still
being thrown properly.

[1] https://docs.ansible.com/ansible/latest/modules/slurp_module.html#examples

Change-Id: I68cd1d7558992029ba9ceca119901410fc9764d5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>