ansible: Fix cleanup for multiple threads
When we switched to single pytest session in [1], it turned out
that one of the fixtures ('ansible_clean_private_dirs') was instantiated
for more than one thread.

The code of the fixture removes ansible's private directories and
raises if the directory in question doesn't exist. For first thread
running the cleanup code everything was fine, but the next one was
trying to remove same dirs again, so an error was raised.

The problem was here probably all the time, but with
multiple pytest sessions it didn't manifest itself.

This patch makes sure that after first cleanup the set containing
directory names is emptied, so that no errors will occur if another
thread tries to run this code.

[1] https://gerrit.ovirt.org/#/c/111441/

Change-Id: I00741d53104ae52f63b167532131976da3ce7d13
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>