basic: Introduce 'vm_ssh' fixture
This patch adds a 'vm_ssh' fixture that allows running
commands on VMs (not the engine/host VMs, but the nested ones).
The difference between the fixture and the original
'_vm_ssh' function is that the latter uses lago-based 'prefix'.

All cases of direct usage of '_vm_ssh' have been replaced
with the new fixture. The remaining ones will be removed
in subsequent patches.

Change-Id: I34b1d087b0b25601ae9d6e39a873d0819d5e5021
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>