basic_sanity: change assert_vm_is_alive to long
When running the OST nested (on a VM), operations are slower that usual.
Therefore, in most cases, the "run_vms" test fails while executing
assert_vm_is_alive(VM0_NAME), with an error of:

../basic-suite-master/test-scenarios/004_basic_sanity.py:196: in is_alive
    _ping, allowed_exceptions=[ansible.AnsibleExecutionError]
../ost_utils/ost_utils/assertions.py:93: in assert_true_within_short
    error_message=error_message)
../ost_utils/ost_utils/assertions.py:75: in assert_equals_within_short
    error_message=error_message

AssertionError: _ping failed to evaluate after 180 seconds.

Changing the assert to long instead of short fixes that issue.

Change-Id: Ib62bd789d294233143a191f30482a525204d2c14
Signed-off-by: Shani Leviim <sleviim@redhat.com>